+++ RELEASE_NOTES.md
@@ -14,6 +14,7 @@
 * Fixed bug (present in 3.5.8) where `ArgumentNullException` is thrown from `INotifyDataErrorInfo.GetErrors` when given `null`
 * Now logging and returning false to WPF if selection in a SubModelSelectedItem binding fails
 * Added prebuilt bindings for `Selector.SelectedIndex`
+* Added Binding.TwoWay.id to API
 
 #### 4.0.0-beta-41
 * Fixed "backwards typing" (#373) and other bugs (like #371) introduced in 4.0.0-beta-40 with more careful use of the Dispatcher (#374)